requirements.txt

flask
pandas
openpyxl
python_decouple
pyodbc